cable_tunnel: use required-features
